+++ nucliadb_fields_tantivy/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "nucliadb_fields_tantivy"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tantivy = "0.17.0"
+nucliadb_service_interface = { path = "../nucliadb_service_interface" }
