+++ nucliadb_node/Cargo.toml
@@ -64,6 +64,7 @@ async-stream = "0.3.2"
 rand = "0.8.4"
 
 # nucliadb dependencies
+nucliadb_services = {path = "../nucliadb_services"}
 nucliadb_vectors = { path = "../nucliadb_vectors" }
 nucliadb_protos = { path = "../nucliadb_protos/rust" }
 nucliadb-cluster = { path = "../nucliadb_cluster"}
