+++ nucliadb_paragraphs_tantivy/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "nucliadb_paragraphs_tantivy"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+lazy_static = "1.4.0"
+regex = "1.5.5"
+tantivy = "0.17.0"
+nucliadb_service_interface = { path = "../nucliadb_service_interface" }
