+++ nucliadb_service_interface/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "nucliadb_service_interface"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+tracing = "0.1.29"
+tempdir = "0.3.7"
+async-trait = "0.1.51"
+async-std = "1.10.0"
+prost-types = "0.10"
+tokio = { version = "1.12.0", features = ["rt-multi-thread", "macros", "sync", "time", "signal", "fs"] }
+nucliadb_protos = { path = "../nucliadb_protos/rust" }
