+++ nucliadb_service_interface/src/lib.rs
@@ -17,18 +17,26 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 //
+pub mod fields_interface;
+pub mod paragraphs_interface;
+pub mod service_interface;
+pub mod vectos_interface;
 
-mod heuristics;
-mod index;
-mod memory_system;
-mod query;
-mod query_delete;
-mod query_insert;
-mod query_post_search;
-mod query_search;
-pub mod reader;
-pub mod service;
-#[cfg(test)]
-mod tests;
-mod utils;
-pub mod writer;
+pub mod dependencies {
+    pub extern crate anyhow;
+    pub extern crate async_std;
+    pub extern crate async_trait;
+    pub extern crate nucliadb_protos;
+    pub extern crate prost_types;
+    pub extern crate tempdir;
+    pub extern crate tokio;
+    pub extern crate tracing;
+}
+
+pub mod prelude {
+    pub use crate::dependencies::*;
+    pub use crate::fields_interface::*;
+    pub use crate::paragraphs_interface::*;
+    pub use crate::service_interface::*;
+    pub use crate::vectos_interface::*;
+}
