+++ nucliadb_service_interface/src/vectos_interface/mod.rs
@@ -17,18 +17,23 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 //
+use crate::service_interface::InternalError;
 
-mod heuristics;
-mod index;
-mod memory_system;
-mod query;
-mod query_delete;
-mod query_insert;
-mod query_post_search;
-mod query_search;
-pub mod reader;
-pub mod service;
-#[cfg(test)]
-mod tests;
-mod utils;
-pub mod writer;
+#[derive(Clone)]
+pub struct VectorServiceConfiguration {
+    pub no_results: Option<usize>,
+    pub path: String,
+}
+
+#[derive(Debug)]
+pub struct VectorError {
+    pub msg: String,
+}
+
+impl std::fmt::Display for VectorError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", &self.msg)
+    }
+}
+
+impl InternalError for VectorError {}
