+++ nucliadb_vectors/Cargo.toml
@@ -20,5 +20,6 @@ tracing = "0.1.29"
 rayon = "1.5.2"
 memmap2 = "0.5.3"
 fs2 = "0.4.3"
+nucliadb_service_interface = { path = "../nucliadb_service_interface" }
 [dev-dependencies]
 env_logger = "0.9.0"
